header.c: bug corrected address.c mime.c misc.c: Always cast to (unsigned char) when using the ctype functions like isspace(); mbx_default.c: If _PATH_SENDMAIL not set via paths.h use "/usr/spool/mail" message.c: uidl is now <md5 . time . sequene> tcp.c: INADDR_NONE is not define on solaris define it to -1.

+++ mailbox/locker.c
@@ -142,7 +142,7 @@ locker_lock (locker_t lock, int flags)
 		 O_WRONLY | O_CREAT | O_EXCL, LOCKFILE_ATTR)) == -1)
     return errno;
   /* Success.  */
-  sprintf(buf, "%d", getpid());
+  sprintf(buf, "%ld", (long)getpid());
   write(fd, buf, strlen(buf));
 
   /* Try to get a file lock.  */
